+++ docs/source/changelog.rst
@@ -15,6 +15,7 @@ Added
 * Support for |Results|.
 * ``pool`` parameter to the |MemcachedRLBackend| rate limiter backend.
 * ``client`` parameter to the |RedisRLBackend| rate limiter backend.
+* ``--watch-use-polling`` command line argument.
 
 Fixed
 ^^^^^